Change API call SplitHostPort slightly
Instead of taking an int for the port, take a short.
This matches better with the networkmanager EndPoint struct.

+++ hub-qt/rpcconsole.cpp
@@ -841,7 +841,7 @@ void RPCConsole::banSelectedNode(int bantime)
     if (CNode *bannedNode = FindNode(strNode.toStdString())) {
         std::string nStr = strNode.toStdString();
         std::string addr;
-        int port = 0;
+        uint16_t port = 0;
         SplitHostPort(nStr, port, addr);
 
         CNode::Ban(CNetAddr(addr), BanReasonManuallyAdded, bantime);
