+++ libs/utils/netbase.h
@@ -223,7 +223,7 @@ public:
 
 CNetAddr::Network ParseNetwork(std::string net);
 std::string GetNetworkName(CNetAddr::Network net);
-void SplitHostPort(std::string in, int &portOut, std::string &hostOut);
+void SplitHostPort(std::string in, uint16_t &portOut, std::string &hostOut);
 bool SetProxy(CNetAddr::Network net, const proxyType &addrProxy);
 bool GetProxy(CNetAddr::Network net, proxyType &proxyInfoOut);
 bool IsProxy(const CNetAddr &addr);
